+++ R/input-checkboxgroupbuttons.R
@@ -49,16 +49,10 @@ checkboxGroupButtons <- function(inputId,
   if (size != "normal") {
     divClass <- paste0(divClass, " btn-group-", size)
   }
-  checkboxGroupButtonsTag <- tags$div(
+  tags$div(
     class = "form-group shiny-input-container shiny-input-checkboxgroup shiny-input-container-inline",
     style = css(width = validateCssUnit(width)),
-    tags$label(
-      id = paste0(inputId, "-label"),
-      class = "control-label",
-      `for` = inputId,
-      label,
-      class = if (is.null(label)) "shiny-label-null",
-    ),
+    label_input(inputId, label),
     if (!is.null(label)) tags$br(),
     tags$div(
       id = inputId,
@@ -87,9 +81,9 @@ checkboxGroupButtons <- function(inputId,
           )
         })
       )
-    )
+    ),
+    html_dependency_buttons_group()
   )
-  attachShinyWidgetsDep(checkboxGroupButtonsTag)
 }