+++ R/input-numericRange.R
@@ -97,13 +97,8 @@ numericRangeInput <- function(inputId,
   rangeTag <- tags$div(
       id = inputId,
       class = "shiny-numeric-range-input form-group shiny-input-container",
-      style = if (!is.null(width)) paste0("width: ", htmltools::validateCssUnit(width), ";"),
-      tags$label(
-        class = "control-label",
-        `for` = inputId,
-        label,
-        class = if (is.null(label)) "shiny-label-null"
-      ),
+      style = css(width = validateCssUnit(width)),
+      label_input(inputId, label),
       tags$div(
         class = "input-numeric-range input-group",
         fromTag,